create dropdown component
